Add a dedicated 'updater' docker image
Until now, the concourse pipeline has been using
the main 'cloud-platform tools' docker image, and
installing any required packages on top of that.

This is overkill, because the tools image has lots
of software which is not required.

This change adds a dedicated image with the
minimal software installed to update the JSON data
in the app. This should enable the concourse
pipeline to be simplified, in a later change.