Accept helm whatup data raw or cooked
We want a JSON document with "apps" and 
"updated_at" keys, which is what all the other
POSTs accept. But we currently get a raw JSON
list (the output of `helm whatup`).

This change allows us to accept either format, so
that the app. will continue to work if it's
updated before the updater app.

Later, when we're confident that the new updater
has been deployed, we can remove this logic.