Add a build step for the updater image
This should push both the web app and updater docker images to docker
hub, for every new release.

This means we will no longer have different tags for the web app and the
updater, but that's not necessarily a bad thing, and trying to change
the logic to have independent tags for each is going to be more trouble
than it's worth.