Parallelize the HOODAW updater
The HOODAW updater pipeline currently takes around 7 minutes to run.
This is tolerable, but as we add more and more reports, it is only going
to get slower.

This change parallelizes the updater, so that the different report tasks
run simultaneously. The `wait` statement means the shell script will not
exit until the slowest reporting task has finished.

This should speed up the pipeline, and hopefully ensure it stays
reasonably quick as we continue to add reports.