Fix `bundle install` problem
There seems to be a problem with the current
version of bundler. AFAICT, it's partially-
ignoring the
```
bundle config set without 'development'
```
command. So, It's specifying that, e.g. rspec and
pry-byebug are *supposed to be* installed (which 
they aren't, for the docker image that runs on
live-1), but not actually installing those gems on
the image.

So, when app.rb is executed, and tries to run 
`bundler/setup`, it fails because the development
gems aren't installed.

I'm sure there is a smarter way to fix this, but
for now I'm just removing the development group
from the gemfile, and installing rspec in the
docker image, even though it's not needed for the
web application - only for the tests.

I tried to do the same thing with pry-byebug, but
I think there are more dependencies required, 
because `bundle install` fails when I include that
gem in the gemfile. So, I'm just leaving it out.