Set page titles dynamically
This change reclaims some of the space on the page
by setting the page title so that it appears in
the h1 tag in the page header, rather than having
a constant "How out of date are we?" in that 
position.

The downside of this change is that we no longer
get `list.size` in the page title, but I think the
trade-off is worth it.