Add cronjobs as a separate, but related helm chart
The cronjobs get deployed into the concourse-main
namespace of the manager cluster, because the 
required secrets are already maintained there.
This way, we don't have to create a second copy of
e.g. our AWS credentials or github admin token.

The cronjobs require some values which are defined
in the main chart, so the makefile supplies some
of the main chart's values when deploying them.

There is probably a much nicer, more helmy way of
doing this, but I don't know what it is yet.