Hardcode the hostname in values-dev.yaml
I'm not sure about this one. I'll probably have a
deployment mechanism to deploy/update both the
web app. and cronjob helm charts at the same time,
so it makes sense to supply values from one to the
other. But, this interpolation doesn't seem to be
working as I expected. So, I'm going to keep them
separate for now.