Remove deprecated stable helm charts repo
Without this line, the helm whatup report shows
that our concourse chart in live-1 is out of date.
With it, the concourse chart is not listed at all.

A lot of charts are not listed, because of errors
like this:
```
WARNING: No Repo was found which containing the Chart 'oauth2-proxy' 
```

Hopefully, these errors will go away as more 
charts are migrated to other repos.